requirements.txt
flask==2.0.3
requests==2.27.1
flask-cors==3.0.10
gunicorn==20.1.0
werkzeug==2.0.3
python-dotenv==0.21.1
colorama==0.4.6
pyautogui==0.9.54
Pillow==9.5.0
